Renaming unit test classes for forms.
