+++ tests/src/Unit/Form/CurrencyImportFormTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Drupal\Tests\currency\Unit\Controller {
+namespace Drupal\Tests\currency\Unit\Form {
 
   use Drupal\Core\Form\FormStateInterface;
   use Drupal\Core\Url;
